Typo
The tagline: "differential DSL that can run fater than light and go back to the past." also has a typo, assuming "faster", but not sure how to fix.

Not sure if hyphen is wanted; also for pre- and post-condition. Auto-diff?

I look forward to looking more into this intriguing package.